+++ .detoxrc.js
@@ -19,7 +19,7 @@ module.exports = {
     simulator: {
       type: 'ios.simulator',
       device: {
-        type: 'iPhone 12'
+        type: 'iPhone 14 Pro'
       }
     },
     attached: {