+++ App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { StatusBar } from 'expo-status-bar';
 
+import { useCachedResources } from 'hooks/useCachedResources';
 import { AppNavigator } from 'navigations';
 
 import ErrorBoundary from 'components/errorBoundary';
@@ -13,6 +14,11 @@ void (async () => {
 })();
 
 export default function App() {
+  const isLoadingComplete = useCachedResources();
+
+  if (!isLoadingComplete) {
+    return null;
+  }
   return (
     <ErrorBoundary>
       <SafeAreaProvider>